@@ -98,64 +97,42 @@
         "babel-plugin-syntax-jsx": "^6.18.0",
         "babel-plugin-transform-remove-strict-mode-tags": "^0.0.1",
         "babel-plugin-transform-vue-jsx": "^3.7.0",
+        "chalk": "^2.0.1",
         "chokidar": "^3.0.2",
         "clean-webpack-plugin": "^3.0.0",
         "compression": "^1.7.4",
-        "cross-env": "^6.0.0",
-        "css-loader": "^3.2.0",
-        "eslint": "^6.1.0",
-        "eslint-loader": "^2.2.1",
-        "eslint-plugin-vue": "^5.2.3",
-        "extract-text-webpack-plugin": "^4.0.0-beta.0",
-        "file-loader": "^4.2.0",
-        "less": "^3.9.0",
-        "less-loader": "^5.0.0",
-        "memory-fs": "^0.4.1",
-        "mini-css-extract-plugin": "^0.8.0",
-        "optimize-css-assets-webpack-plugin": "^5.0.3",
-        "ora": "^4.0.3",
-        "postcss-loader": "^3.0.0",
-        "postcss-preset-env": "^6.7.0",
-        "postcss-px2rem": "^0.3.0",
-        "prettier": "^1.19.1",
-        "rimraf": "^3.0.0",
-        "style-loader": "^1.0.0",
-        "stylelint": "^10.1.0",
-        "stylelint-webpack-plugin": "^0.10.5",
-        "uglifyjs-webpack-plugin": "^2.2.0",
-        "url-loader": "^2.1.0",
-        "vue-loader": "^15.7.1",
-        "vue-style-loader": "^4.1.2",
-        "vue-template-compiler": "^2.6.10",
-        "webpack-cli": "^3.3.6",
-        "webpack-dev-middleware": "^3.7.0",
-        "webpack-hot-middleware": "^2.25.0",
-        "webpack-merge": "^4.2.1",
-        "webpack-node-externals": "^1.7.2",
-        "chalk": "^2.0.1",
-        "commitizen": "^2.9.6",
-        "commitlint-config-cz": "^0.12.1",
         "compression-webpack-plugin": "^3.0.1",
+        "cookie-parser": "~1.4.4",
         "copy-webpack-plugin": "^4.0.1",
         "core-js": "^2.6.11",
+        "cross-env": "^6.0.0",
+        "css-loader": "^3.2.0",
         "cssnano": "^4.1.10",
-        "cz-customizable": "^6.2.0",
+        "debug": "~2.6.9",
+        "ejs": "~2.6.1",
         "ejs-babel-loader": "^1.3.7",
+        "eslint": "^6.8.0",
         "eslint-friendly-formatter": "^4.0.1",
+        "eslint-loader": "^2.2.1",
         "eslint-plugin-html": "^6.0.0",
         "eslint-plugin-import": "^2.19.1",
         "eslint-plugin-node": "^10.0.0",
         "eslint-plugin-prettier": "^3.1.2",
         "eslint-plugin-promise": "^4.2.1",
+        "eslint-plugin-vue": "^5.2.3",
         "express": "^4.17.1",
+        "extract-text-webpack-plugin": "^4.0.0-beta.0",
+        "file-loader": "^4.2.0",
         "friendly-errors-webpack-plugin": "^1.6.1",
         "happypack": "^5.0.0",
         "hash-sum": "^2.0.0",
         "highlight.js": "^9.17.1",
         "html-loader": "^0.5.5",
         "html-webpack-plugin": "3.2.0",
-        "husky": "^0.15.0-beta.16",
+        "http-errors": "~1.6.3",
         "json5": "^2.1.1",
+        "less": "^3.9.0",
+        "less-loader": "^5.0.0",
         "lint-staged": "^6.0.0",
         "lodash": "^4.17.15",
         "lru-cache": "^5.1.1",
