@@ -3,7 +3,7 @@
         <div class="home-images">
             <img class="home-images-cover1" src="/public/SaoirseRonan1.jpg" />
         </div>
-        <router-link class="home-link" to="/detail">MORE!</router-link>
+        <router-link class="home-link" to="/detail">MORE</router-link>
     </div>
 </template>
 
